### 2019-07-23 #1575 @gus0k
I updated the package to version 0.7.5, after a small error in how the submodules were being loaded. How should I change that?
