### 2019-07-23 #1591 @whedon
[REVIEW]: PyMarket
**Submitting author:** @gus0k (<a href="http://orcid.org/0000-0001-8041-9685">Diego</a>)
**Repository:** <a href="https://github.com/gus0k/pymarket" target ="_blank">https://github.com/gus0k/pymarket</a>
**Version:** v0.7.8

    -  ~gus0k/pymarket/issues/13~

    -  gus0k/pymarket/issues/4

    - ~gus0k/pymarket/issues/3~
    - gus0k/pymarket/issues/5
    - gus0k/pymarket/issues/7

    - ~gus0k/pymarket/issues/3~
    -  ~gus0k/pymarket/issues/2~
