### 2019-07-31 #1591 @gus0k
Hello,

Do you have anything so far that you would like me to fix or clarify?

Best regards,
K
