### 2019-08-05 #1591 @igarizio
Hello @gus0k 👋,
Great work on PyMarket! It is really well written and nicely documented. Also, the idea behind it seems really useful.

Here are some of my comments:

- Software version: This may be silly, but I prefer to ask. The review states v0.7.4 but your last version is v0.7.5, I suppose you want to publish the last version, right? (@alexhanna if that is the case, is it possible to just change the version stated in this issue?)

- Authorship: Just wanted to check, but you own both @gus0k and @ellguso, right? (context: I was looking at the [contributors](https://github.com/gus0k/pymarket/graphs/contributors) part of your repo)

- Paper acknowledgement: Here I may be wrong, but I think the title "acknowledgement" may not be correct. What I generally see in the acknowledgement part is words of appreciation to other researchers, funding entities, etc. (@taqtiqa-mark, @alexhanna please let me know what you think).

- List of mechanisms: I think it would help people reading your paper if you added a list of the mechanisms implemented in the library (maybe you can even add this list to your README).

- Paper references: If you decide to add the list of mechanisms implemented, I think it would be appropriate to cite where those mechanisms come from (I think you already have those references inside huang_auction.py, muda_auction.py, etc.)

- Small detail: On the second page of the paper, it says "Figure 1: png".

I will continue to review the other items in the checklist and will post if necessary. Please let me know if you have any questions. 😃
