### 2019-08-06 gus0k/pymarket#2 @taqtiqa-mark
List algorithms and their authors in JOSS paper.
Algorithms implemented and their references should be cited (simple list) in JOSS paper.

Related: openjournals/joss-reviews/issues/1591
