### 2019-08-06 gus0k/pymarket#3 @taqtiqa-mark
JOSS paper statement of need: Algorithms implemented
JOSS paper statement of need should cite markets algorithms implemented.
