### 2019-08-06 gus0k/pymarket#4 @taqtiqa-mark
JOSS paper: Synchronize version numbers
Synchronize version numbers such that the JOSS release version given match the GitHub release.
