### 2019-08-08 gus0k/pymarket#3 @gus0k
Fixed in commit https://github.com/gus0k/pymarket/commit/2f8db92010d5f9407a72941788500351e92cbe81
