### 2019-08-08 #1591 @gus0k
@igarizio @taqtiqa-mark Thank you both so far for your input. I have updated the paper to list the implemented algorithms and their references.

@alexhanna  Would it be possible to update the version number to 0.7.5?

Regarding the acknowledgements, Telecom ParisTech is paying for my PhD. In that sense, I feel it plays the same role as funding entities.

Best regards,
K
