### 2019-08-30 gus0k/pymarket#7 @taqtiqa-mark
Requirements specification incorrect: Matplotlib
### Description

To install pip packages the requirements file suggests there are no version restrictions.
A run of Travis CI clearly warns [at least one test will break for Matplotlib >=3.3](https://travis-ci.org/gus0k/pymarket/builds/578659316#L453)
