### 2019-09-03 gus0k/pymarket#7 @gus0k
The dependency showing the error is not used, so that can be fixed easily.
I am not sure how I would go about to check the minimum version required for each library. Would you need me to do so?
