### 2019-09-03 gus0k/pymarket#5 @gus0k
I do not fully understanad what you mean by this. Dependencies are listed in requirements.txt and are automatically installed if the package is installed using pip.
I can add to the read-the-docs if that is what you are asking.
