### 2019-09-03 gus0k/pymarket#7 @taqtiqa-mark
Isn't there is a minimum python version?  this should be in the requirements file - I believe python allows this.

For package minimum versions take the smallest version numbers that you know to work the time of publication.

**Yes please add minimum versions numbers** - the role of your users is not to discover minimum package versions by discovering/reporting what versions do not work.
