### 2019-09-04 gus0k/pymarket#7 @gus0k
I added "minimum" versions in the requeriments.txt file as well as in the setup.py
