### 2019-09-04 gus0k/pymarket#5 @gus0k
I updated in the installation section a comment regarding the lack of external dependencies. I think that referring to the requeriments.txt should be enough regarding pure python dependencies.
