### 2019-09-24 #1591 @alexhanna
Hi @taqtiqa-mark, are you able to continue reviewing this?

👋 @igarizio, it looks like the last thing you need to check on the review is the functionality?
