### 2019-10-07 gus0k/pymarket#11 @taqtiqa-mark
Instructions incomplete on Ubuntu 16.04
* pymarket version: pip installed version (version 0.7.5 ?)
* Python version: python3-pip=8.1.1-2ubuntu0.4
* Operating System: Ubuntu 16.04

### Description

Trying to install pymarket following the instructions.
The instructions result in an error.

Example of an instance of issues #5 and #8.

### What I Did

On Ubuntu 16.04, with latest packaged pip installed....
```
apt install --yes python3-pip=8.1.1-2ubuntu0.4
```
 the instructions abend:
```
$ pip3 install pymarket
Collecting pymarket
    default:   Downloading https://files.pythonhosted.org/packages/84/53/d74deabe429cbd7cf809b6a272302db251d74c06509f74b86d0f4d1899f2/pymarket-0.7.5-py2.py3-none-any.whl
    default: Collecting pandas>=0.24 (from pymarket)
    default:   Downloading https://files.pythonhosted.org/packages/d9/05/38875a81040e679c196a854865dbafe4dfe5f92e8365ddfff21f2817d89d/pandas-0.25.1-cp35-cp35m-manylinux1_x86_64.whl (10.3MB)
    default: Collecting pulp (from pymarket)
    default:   Downloading https://files.pythonhosted.org/packages/2d/33/3ae6d9d2ac8c7068937af6372fd8828ac605e62a8b17106fe57110930d38/PuLP-1.6.10.zip (13.6MB)
    default: Collecting numpy>=1.16 (from pymarket)
    default:   Downloading https://files.pythonhosted.org/packages/9b/21/2b18339d24a2f73dcefb2f10f48aff6182e16da83e3a612684443c6cfb29/numpy-1.17.2-cp35-cp35m-manylinux1_x86_64.whl (20.1MB)
    default: Collecting networkx (from pymarket)
    default:   Downloading https://files.pythonhosted.org/packages/85/08/f20aef11d4c343b557e5de6b9548761811eb16e438cee3d32b1c66c8566b/networkx-2.3.zip (1.7MB)
    default: Collecting Click>=6.0 (from pymarket)
    default:   Downloading https://files.pythonhosted.org/packages/fa/37/45185cb5abbc30d7257104c434fe0b07e5a195a6847506c074527aa599ec/Click-7.0-py2.py3-none-any.whl (81kB)
    default: Collecting matplotlib (from pymarket)
    default:   Downloading https://files.pythonhosted.org/packages/12/d1/7b12cd79c791348cb0c78ce6e7d16bd72992f13c9f1e8e43d2725a6d8adf/matplotlib-3.1.1.tar.gz (37.8MB)
    default:     Complete output from command python setup.py egg_info:
    default:
    default:     Beginning with Matplotlib 3.1, Python 3.6 or above is required.
    default:
    default:     This may be due to an out of date pip.
    default:
    default:     Make sure you have pip >= 9.0.1.
    default:
    default:
    default:     ----------------------------------------
```
