### 2019-10-07 gus0k/pymarket#11 @taqtiqa-mark
Steps to reproduce (using a Digital Ocean VM):
```
push dev
vagrant up
vagrant provision
```
