### 2019-10-12 gus0k/pymarket#13 @igarizio
Error when generating uniform bidders
* pymarket version: 0.7.5
* Python version: 3.6.7
* Operating System: Windows 10

### Description

Generating uniform bidders with a number of buyers or sellers greater than 1/eps returns a ValueError.

```python
<ipython-input-48-62e55ce4d0c0> in <module>
      2
      3 mar = pm.Market()
----> 4 bids = pm.datasets.uniform_bidders.generate(10001, 10001, 10, 10)
      5 for b in bids:
      6     mar.accept_bid(*b)

~\Anaconda3\lib\site-packages\pymarket\datasets\uniform_bidders.py in generate(cant_buyers, cant_sellers, offset_sellers, offset_buyers, r, eps)
     55     for i, (o_, q_) in enumerate(zip(offset, quantities)):
     56         range_ = np.arange(0, 1, eps)
---> 57         qs = r.choice(range_, q_, replace=False)
     58         vs = r.choice(range_ + o_, q_, replace=False)
     59         for j in range(q_):

mtrand.pyx in mtrand.RandomState.choice()

ValueError: Cannot take a larger sample than population when 'replace=False'
```
I think this can be solved just by setting eps accordingly.

### What I Did

```python
pm.datasets.uniform_bidders.generate(10001, 10001, 10, 10)
```
