### 2019-10-13 #1591 @alexhanna
Hi @taqtiqa-mark, checking in again. Can you confirm that the issues noted above have been adequately addressed?
