### 2019-10-13 #1591 @taqtiqa-mark
No. I've opened gus0k/pymarket/issues/11 as an example of issues 5/7, and I've made a pull request  providing reproducible example.  I'll settle for just Ubuntu 16.04 instructions that work.
