### 2019-10-14 gus0k/pymarket#11 @gus0k
I think the problem you are having is due to the fact that in Ubuntu 16.04, python3 is actualy 3.5 and 3.6.

I created a docker to reproduce this problem. I'm have no experience with Vagrant, but i believe this should be similar enough.
The new dockerfile is Dockerfile_xenial created in https://github.com/gus0k/pymarket/commit/5241d054b33705ec519393b89d4416285b931c1d

To test it , run:
```sh
docker build -f Dockerfile_xenial -t pymarket:xenial .
 docker run -it pymarket:xenial
```
Once in the console, you can open open python using:
```sh
python3
```
I used one of the examples in the the documentation and it works fine:

```python

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import networkx as nx
import pymarket as pm
r = np.random.RandomState(1234)
mar = pm.Market()

mar.accept_bid(1, 6.7, 0, True, 0)
mar.accept_bid(1, 6.6, 1, True, 0)
mar.accept_bid(1, 6.5, 2, True, 0)
mar.accept_bid(1, 6.4, 3, True, 0)
mar.accept_bid(1, 6.3, 4, True, 0)
mar.accept_bid(1, 6, 5, True, 0)

mar.accept_bid(1, 1, 6, False, 0)
mar.accept_bid(1, 2, 7, False, 0)
mar.accept_bid(2, 3, 8, False, 0)
mar.accept_bid(2, 4, 9, False, 0)
mar.accept_bid(1, 6.1, 10, False, 0)

bids = mar.bm.get_df()
transactions, extras = mar.run('p2p', r=r)
stats = mar.statistics()
```
