### 2019-10-14 #1591 @gus0k
@taqtiqa-mark Sorry for the dealy, it was a busy week. I have already addressed the issue. I think there was a problem with the python version you were using. I uploaded a working Docker with ubuntu 16.04

@igarizio Fixed the problem with too many bidders.
