### 2019-10-15 gus0k/pymarket#11 @taqtiqa-mark
This should not be closed:

1. The output indicates it is a pip issue.
1. The Travis config suggests Xenial is supported.
