### 2019-10-15 gus0k/pymarket#11 @gus0k
The docker i uploaded is running with an old version of pip (8) and it works flawlessly:

```python
Python 3.6.8 (default, Oct  9 2019, 14:04:01)
[GCC 5.4.0 20160609] on linux
>>>
KeyboardInterrupt
>>>
```
```sh
root@bdf6686310c3:/# python3 -m pip --version
pip 8.1.1 from /usr/lib/python3/dist-packages (python 3.6)
```

If you want me too add a line that says that a recent pip is recommended, I'll do it. But it is a python's version problem.
