### 2019-10-15 gus0k/pymarket#13 @igarizio
Hi! The commit introduced a new bug. The new code should be bellow:
 ```python
quantities = [cant_buyers, cant_sellers]
```
Also, I think it would be a good idea to clarify in the doctring the change in eps when there are too many buyers/sellers.
