### 2019-10-15 gus0k/pymarket#11 @taqtiqa-mark
I just want instructions that anyone can follow and get up and running.

You need to add update pip instructions (see your .travis.yml file) if you want to say Ubuntu Xenial works out if the box.

It shouldn't be necessary to have docker or vagrant to get this running.
