### 2019-10-15 gus0k/pymarket#13 @gus0k
Done: pymarket/datasets/uniform_bidders.py
