### 2019-10-23 #1591 @alexhanna
Hi @taqtiqa-mark can you ensure that these issues have been addressed and check off the relevant boxes in your review?
