### 2019-10-24 gus0k/pymarket#15 @taqtiqa-mark
Blocking: Tested working install instructions.
* pymarket version: latest
* Python version: 3.6
* Operating System: Ubuntu Bionic

### Description

I followed the install instructions.

However, after upgrading pip via `pip install --user --upgrade pip` pip refuses to start and raises this instead:

```bash
Traceback (most recent call last):
  File "/usr/bin/pip3", line 9, in <module>
    from pip import main
ImportError: cannot import name 'main'
```

This apparently a well known thing, previously reported.
- https://stackoverflow.com/questions/49836676/error-after-upgrading-pip-cannot-import-name-main
- https://stackoverflow.com/questions/28210269/importerror-cannot-import-name-main-when-running-pip-version-command-in-windo
- https://askubuntu.com/questions/1108070/pip3-command-fails-after-upgrading-to-pip-version-18-1
- https://askubuntu.com/questions/1135252/pip-broken-after-upgrading-to-19-04
- https://askubuntu.com/questions/1025189/pip-is-not-working-importerror-no-module-named-pip-internal
- https://askubuntu.com/questions/1025793/running-pip3-importerror-cannot-import-name-main

**I hope you can appreciate how frustrating this is?  You have broken a users pip installation on their computer!!**

Given that you appear to refuse to test your instructions and just wave your hand that this works everywhere:

The only way I can close this issue is:
1. Update the vagrant scripts such that they work.  then update your install instructions accordingly.
2. Remove all references to install instructions for Linux, and state the only way to get this to run is to use the docker container you provide, and give the link to the built container.

As I commented [here](https://github.com/openjournals/joss-reviews/issues/1591#issuecomment-545712658), the fastest way to publish this package is the second of the above options.

### What I Did

```bash
git clone https://github.com/gus0k/pymarket
pushd pymarket/dev/bionic
vagrant up
```
