### 2019-10-24 #1591 @taqtiqa-mark
@alexhanna As best I can tell the install instructions are not tested.
@gus0k previously waved his hands that things work on Linux, but has now qualified that to a couple of distributions.  The install instructions have been updated to add some missing step. So we are making progress (painfully).

I've made another pull request, showing the install instructions will break the users pip installation on their computer - a big no-no.

Given the package is currently so small and has limited use cases, I have suggested to @gus0k that simply providing a link to a working Docker Container, and removing references to claims to work on Linux is the fastest path forward.
