### 2019-10-28 gus0k/pymarket#15 @gus0k
@taqtiqa-mark I'm sorry about the error in the installation instructions. I haven't been able to set up the digital ocean account (which as far as i understand are the images your are using), so I could not reproduce your results exactly.

I understood what was pushing the library to python version 3.6> and I'm gonna try to make some changes in order to make it as easy to install as possible. If this fails, we can resort to using your Docker approach.

I should have something for you in a couple of days. I'll update you as soon as possible.
