### 2019-10-28 gus0k/pymarket#15 @gus0k
@taqtiqa-mark

I made some changes that might could solve your installation issues. The build is working on travis but my internet connection is pretty bad and I won't be able download new dockers to test it there.

If you could try the following, it can help me understand if this changes are in the right direction:

Without updating pip or anything, try to run the following (change python to python3 if you have both available), or build from the branch "python 3.5":

```sh
python -m pip install https://github.com/gus0k/pymarket/archive/python3.5.zip --user
```

Basically, I made the library compatible with python 3.5 and decreased the minimum requirements it has. I think this should suffice to solve the installation issues as it runs on the default configuration of most Ubuntus.

Looking forward to hear from you,
Diego
