### 2019-10-28 gus0k/pymarket#15 @taqtiqa-mark
@gus0k Thanks for making these changes.  A Digital Ocean account is not necessary: VirtualBox or VM Ware will allow you to obtain a fresh Linux installation (whichever you think you can get working) to start working from.

Please either install some VM package and **_you_** develop and test **_your_** install instructions.

Alternatively make the docker image that you claim works available, and remove references to linux installations and instructions.
