### 2019-10-30 gus0k/pymarket#15 @gus0k
@taqtiqa-mark I tested it on a local vagrant instalation and it's working.

In branch python3.5, inside dev/diego you can find to vagrantfiles, one for bionic and one for xenial (one is running python3.5 and the other python 3.6).

The code I ran is the following

```sh
vagrant up
vagrant provision
vagrant reload
vagrant ssh
python3 -m pip install https://github.com/gus0k/pymarket/archive/python3.5.zip --user
```
The last line I ran it inside the vagrant because for some reason I wasn't able to install anything (not only pymarket) from the provision line.

It works for both systems. Right now works for python >= 3.5, no need to update pip.
