### 2019-10-30 gus0k/pymarket#15 @taqtiqa-mark
Any reason you don't run the test suite?
Can you pipe the output showing clean runs.

Right now this will only expected/shown to work on Ubuntu Bionic 18.04 correct?
