### 2019-10-31 gus0k/pymarket#15 @gus0k
@taqtiqa-mark , it's done now.

You can take a look at both provision.log files in the corresponding folders.

Right now it has been tested in: my local pc, in all the travis you can check from the main page and two vagrant files:

# dev/diego/bionic

Ubuntu 18.04.03, Python 3.6.8, 9.0.1

# dev/diego/xenial

Ubuntu 16.04.0.1, Python 3.5.2, pip 8.1.1
