### 2019-11-08 gus0k/pymarket#15 @taqtiqa-mark
Can you let me kown when you have merged this into master and updated your instructions to reflect what is in those scripts - in a brief read I noted some package installs that are new.
