### 2019-11-10 gus0k/pymarket#15 @taqtiqa-mark
Hi,  I consulted:
https://pymarket.readthedocs.io/en/master/installation.html

Please do the following:
1. Start a some fresh OS, which ever you say your instructions work for.
2. Read the instructions on the above page
3. As you encounter an instruction,
   a. type it into the console
   b. press enter.

I believe you will discover your instructions are incomplete.
