### 2019-11-10 gus0k/pymarket#15 @taqtiqa-mark
>  I just tried in my friend's PC and it works.

How can that be true?:

![image](https://user-images.githubusercontent.com/1468258/68550779-34586f00-045a-11ea-986f-cc72f9be116b.png)
