### 2019-11-10 gus0k/pymarket#15 @gus0k
That seems to be an old version. You need to access the docs built against master, the other only updates with a version bump, which I haven't done.

https://pymarket.readthedocs.io/en/latest/installation.html
