### 2019-11-12 gus0k/pymarket#15 @taqtiqa-mark
Apologies for the terse reply above.

I think I understand what you've done: You're Vagrant files are the setup for installing from source.
Correct?

If so, your instructions for the "From Sources" (https://pymarket.readthedocs.io/en/latest/installation.html#from-sources) section should be :

For Ubuntu 16.04 & 18.04:
```bash
apt-get install --yes  python3-pip=9.0.1-2.3~ubuntu1.18.04.1
apt-get install --yes pkg-config
apt-get install --yes libfreetype6-dev
apt-get install --yes libpng12-dev
python3 -m pip -V
echo "Installing setupttools"
python3 -m pip install 'setuptools>=27.3'
echo "Checking setuptools"
python3 -m pip freeze | grep setuptools

cd /~/src
git clone -b python3.5 https://github.com/gus0k/pymarket.git
cd pymarket
python3 setup.py install
python3 setup.py test
```

The broken pip installation that I report in this ticket is from following the stable release instructions.
Can you please workout stable release install instructions as detailed as the above.
Once you have worked that out, please add them to the existing Vagrantfile(s) I provided.
