### 2019-11-17 gus0k/pymarket#15 @gus0k
Hello,

Sorry for my late replay. I have been busy with a deadline.
To make it easy to maintain, I erased all documentations but latest version, there is no longer a stable one. In the future, if we have many versions we might consider brining it back up.

Without wanting to sound rude, I have provided two working vagrant files as you requested at some point. Even If I modify your vagrant files, because the base image is different, the needs might be different. Because I cannot run it, I would probably just break it more.

Instructions as they are know work fine for anyone that wants to use the library, and they are a guide (testes in Ubuntu 16/18) on how to run also the tests.

Feel free to try those instructions in your own images. If that does not work, we can try to figure out why it does not for each particualar case.

Best regards,
DIego
