### 2019-11-17 #1591 @gus0k
Hello,

I think I managed to fix the problem: provide working installation instructions and specify for which platforms does it work.
I just addressed @taqtiqa-mark last comment.

Best regards,
Diego
