### 2019-11-18 gus0k/pymarket#15 @taqtiqa-mark
I think you've misunderstood the documentation instructions that this issue relates to.

I am referring to the instructions as at this date, at this URL:
https://pymarket.readthedocs.io/en/latest/installation.html

The Vagrant files I provided tried to get to the point where the following command would run (these instruction is still there as at the time of writing):

`python -m pip install pymarket --user`

In response you provided Vagrant files that get a user to the point that the following command would run:

`python3 setup.py install`
