### 2019-11-18 gus0k/pymarket#15 @gus0k
Have you tried your Vagrant files again with the new version of pymarket? If not, please try that:
I think that they should be working now as they are.
If they do not, please let me know because there might be an issue I have not seen yet.

Best regards,
Diego
