### 2019-11-18 gus0k/pymarket#15 @taqtiqa-mark
If you merge the pull request and run:
```
pushd dev/xenial
vagrant up --provider=virtualbox
```
 and

```
pushd dev/xenial
vagrant up --provider=virtualbox
```

I believe you will see that both are broken in different ways.

I suggest getting just one (bionic?) to work, and remove references to Xenial, Mandriva, etc. until you have added a build I can verify works - sorry I can't any longer accept assurances that things work given the history of those claims.
