### 2019-11-18 gus0k/pymarket#15 @gus0k
As far as I can see from your install.sh files your are running the tests as
`make test`

when the updated version of the installation uses (https://pymarket.readthedocs.io/en/latest/installation.html):
`pytest`

Also, the requeriments_dev.txt is not refered in the documentation and is clearly outdated. I'll remove it. If you want to run the test the way you are doing it you most likely only need to install:

`python3 -m pip install pytest --user`

Regarding Vagrant, for some reason, my computer is not running any vagrant at all. I'm running into several errors which do not seem to be easily fixed.
If you insist on having the tests run in Vagrant (as opposed to Docker), It would help me to have the logs that you obtain.

Best regards,
Diego
