### 2019-11-18 gus0k/pymarket#15 @taqtiqa-mark
> Also, the requeriments_dev.txt is not refered in the documentation and is clearly outdated.

Yes please remove it if you no longer recommend its use.  It is still referenced in your original docker file, here:
https://github.com/gus0k/pymarket/blob/master/dev/Dockerfile#L12
which was the starting point to verify the instructions.

I haven't been able to reproduce your steps using a VM. Aside from installing a fresh OS on my laptop, a VM is the only way to verify your instructions work for some other user.
And the only way to ensure your instructions don't break my work machine - I showed your earlier instructions would have broken my pip installation - happily I dodged that bullet.

Vagrant is the way I've launched a VM, as a courtesy I contributed those scripts.  Unfortunately that didn't work out.
There was, and is, nothing stopping you from testing your instructions on something such as VirtualBox or some such environment.

**Due to non-responses. I'm abandoning the effort to verify your installation instructions in a classical user environment.**

I will now try to verify the docker images build and run.

Please remove any reference to pymarket being able to be installed and run on anything other than docker images.
