### 2019-11-19 gus0k/pymarket#15 @gus0k
I just downloaded Ubuntu 16 and 18 from the official website and created two virtual machines.

After running normal updates: `sudo apt-get update` and installing pip `sudo apt-get install python3-pip`,

I was able to run the installation instructions without any problem:

```sh
python3 -m pip install pymarket
```
to test it:

```sh
git clone https://github.com/gus0k/pymarket
cd pymarket
python3 -m pytest
```

And it worked like a charm.
