### 2019-11-19 gus0k/pymarket#15 @taqtiqa-mark
Please see the updated pull request.  Please make changes to this scripts that reproduce the following failure:

> ... it worked like a charm.

```
<snip>
pymarket1604: Successfully installed cycler-0.10.0 decorator-4.4.1 kiwisolver-1.1.0 matplotl
ib-2.2.4 networkx-2.2 numpy-1.17.4 pandas-0.23.0 pulp-1.6.10 pymarket-0.7.6 pyparsing-2.4.5 pyth
on-dateutil-2.8.1 pytz-2019.3
    pymarket1604: You are using pip version 8.1.1, however version 19.3.1 is available.
    pymarket1604: You should consider upgrading via the 'pip install --upgrade pip' command.
    pymarket1604: Cloning into 'pymarket'...
    pymarket1604: /usr/bin/python3: No module named pytest
```
