### 2019-11-19 gus0k/pymarket#15 @gus0k
It seems that you haven't installed pytest and it seems that I haven't added it in the documentation. I'm truly sorry about that. In the comment above, I meant pytest instead of pymarket in the first line of the testing procedure (typing error)

```sh
python3 -m pip install pytest
```
If you install pymarket from source, then it gets installed automatically (that is what is described in the documentation). If you mix both installations (from pip and then clone), you need to install pytest additionaly (that case is not covered in the documentation yet, but works if you install pytest on its own).
