### 2019-11-19 gus0k/pymarket#15 @taqtiqa-mark
No worries.
Please make the changes you say are needed when you merge the current pull request.

I hope you understand why I won't be able to a close these issues while OS's are cited as working, but for which there is no evidence.  Evidence would be:
 - scripts/instructions I have tested
-  the other reviewer states they have tested OS xyz.
