### 2019-11-20 gus0k/pymarket#15 @gus0k
Hello,

I merged your pull request and added the line missing to install pytest. I also updated the documents to reflect the fact that pytest needs to be installed to run tests.

I think that your test should run now, unless something strange occurs.

Thank you for your time,
Diego
