### 2019-11-25 gus0k/pymarket#15 @gus0k
@taqtiqa-mark Have you had the opportunity to test your vagrant files with my modifications?
