### 2019-12-02 #1591 @taqtiqa-mark
The install instructions now appear to work.  I haven't had time to check the remaning items.  Expect to this weekend.
