### 2019-12-09 #1591 @alexhanna
👋 @taqtiqa-mark - have you been able to check the remaining items?
