### 2020-02-11 #1591 @danielskatz
some suggested changes for the paper are in https://github.com/gus0k/pymarket/pull/18 and for the bib are in https://github.com/gus0k/pymarket/pull/19
