### 2020-02-11 #1591 @danielskatz
Also, please add the 2 missing DOIs above to the .bib file
