### 2020-02-12 #1591 @gus0k
@danielskatz
Created the tagged version (0.78) and uploaded it to zenodo:

10.5281/zenodo.3665175
